Rename : Reorganize directory paths for models and repositories
